+++ koishi/plugins/automation_commands/__init__.py
@@ -1,12 +1,12 @@
 from .commands import *
 from .constants import *
+from .list_all import *
 from .permission_checks import *
-from .show_all import *
 
 
 __all__ = (
     *commands.__all__,
     *constants.__all__,
+    *list_all.__all__,
     *permission_checks.__all__,
-    *show_all.__all__,
 )