+++ koishi/plugins/automation_welcome/constants.py
@@ -1,5 +1,11 @@
 __all__ = ()
 
+from hata import GuildProfileFlag
+
+ONBOARDING_MASK_STARTED = GuildProfileFlag().update_by_keys(onboarding_started = True)
+ONBOARDING_MASK_COMPLETED = GuildProfileFlag().update_by_keys(onboarding_completed = True)
+ONBOARDING_MASK_ALL = ONBOARDING_MASK_STARTED | ONBOARDING_MASK_COMPLETED
+
 WELCOME_MESSAGES = (
     lambda user: f'Roses are red, also blue, {user.mention} joined this server with Mr. Hat too',
     lambda user: f'Koooosh. {user.mention} just landed.',